Add full streaming support and indefinite-length string encoding (#341)
This is a pretty large PR on the encoding side -- should all be OK though, tests are passing and lots of review.


* Streaming support (just a start)

* Checkpoint... more streaming stuff is working

* check point

* UsefulOutBuf test coverage back to 100%; other improvements

* Fix ifdef fan out

* Most API design seems good now

* Missed files from last commit

* More ifdef clean up

* Clean up and documentation

* indefinite length strings

* Fix ifdef fannout

* fix ifdef fannout

* Indefinite length strings and lots other working

* fix ifdef fannout

* More ifdef fan out fix

* Lots of documentation, test coverage...

* Increase test coverage

* Fix #ifdef fanout and documentation clean up

* Usefulbuf flush error is now just an int

* UsefulBuf comments, test coverage and minor improvement

* Final tidy-up of UsefulBuf -- good enough for beta quality

* Fix error codes

* Make QCBORPrivateMajorType private

* Code tidy up

---------

+++ src/qcbor_main_decode.c
@@ -2548,7 +2548,7 @@ MemPool_Pack(UsefulBuf Pool, uint32_t uFreeOffset)
    UsefulOutBuf_Init(&UOB, Pool);
    UsefulOutBuf_AppendUint32(&UOB, (uint32_t)Pool.len); // size of pool
    UsefulOutBuf_AppendUint32(&UOB, uFreeOffset); // first free position
-   return UsefulOutBuf_GetError(&UOB);
+   return (int)UsefulOutBuf_GetError(&UOB);
 }
